refactor: split structured-data template core into focused modules
342 lines and 30 exports covered four unrelated jobs: value primitives,
markdown/lexical text reduction, entity field resolution, and media URL
selection.

Split accordingly, keeping the service file as a barrel so both builder
services (singleTypeListicles, listicleItineraries) keep one import site
(342 -> 56 lines):
- structured-data-primitives.ts       guards, path picking, compaction
- structured-data-text.service.ts     text -> structured description
- structured-data-entity.service.ts   entity fields + price range
- structured-data-media.service.ts    selected photo / IG permalink

Pure extraction. Verified the export surface is identical (30/30 names via
the TS checker, diffed against main).

Verified: tsc error signature identical to baseline; vitest 333 pass / 3
pre-existing failures, same as main — and both dependent
structured-data-template.service tests pass.
